refactor(sheets): extract FloatingCombatBar state to separate file
Move combat state logic, pip interaction handlers, and animation
effects from FloatingCombatBar.svelte into FloatingCombatBar.svelte.ts
to reduce component complexity and improve maintainability.
